phase 3.6: ship agent actions module
